requirements.txt
pytorch-lightning ==1.8.4
PyYAML==6.0
tensorboard>=2.2.0

torchsummary==1.5.1

matplotlib>=3.6.3

#mit a jour des librairies
